fixed wrong api calls & removed dummy methods

+++ case-be/src/main/java/com/eminyilmazz/casebe/controller/LocationController.java
@@ -17,7 +17,8 @@ public class LocationController {
     LocationService locationService;
 
     @GetMapping(value = "/get", produces = MediaType.APPLICATION_JSON_VALUE)
-    public List<PlaceDTO> get(@RequestParam(name = "long") double longitude, @RequestParam(name = "lat") double latitude, @RequestParam(name = "rad", required = false, defaultValue = "1000.0") double rad) {
+    public List<PlaceDTO> get(@RequestParam(name = "long") double longitude, @RequestParam(name = "lat") double latitude, @RequestParam(name = "rad", required = false, defaultValue = "1000.0") int rad) {
+
         return locationService.get(LocationDTO.builder()
                 .longitude(longitude)
                 .latitude(latitude)
