+++ case-be/src/main/java/com/eminyilmazz/casebe/repository/LocationRepository.java
@@ -9,8 +9,8 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface LocationRepository extends JpaRepository<Location, Long> {
     @Query("select (count(l) > 0) from Location l where l.latitude = :lat and l.longitude = :long and l.radius = :rad")
-    boolean existsByLatitudeAndLongitudeAndRadius(@Param("lat") double latitude, @Param("long") double longitude, @Param("rad") double radius);
+    boolean existsByLatitudeAndLongitudeAndRadius(@Param("lat") double latitude, @Param("long") double longitude, @Param("rad") int radius);
 
     @Query("select l from Location l where l.latitude = :lat and l.longitude = :long and l.radius = :rad")
-    Location findByLongitudeAndLongitudeAndRadius(@Param("lat") double latitude, @Param("long") double longitude, @Param("rad") double radius);
+    Location findByLongitudeAndLongitudeAndRadius(@Param("lat") double latitude, @Param("long") double longitude, @Param("rad") int radius);
 }
